+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0-beta4
+
+- Adds the concept of a reponse object
+- Update all callbacks to provide `err, body, response`.
+- Change default returned value for promises from `body` to `response`.
+- Remove `client.batchFeatureUpdate`. `client.insertFeature` or `client.destroyFeture` should be used now.
+
 ## 1.0.0-beta3
 
 - Update tilestat service for new API.